feat: Adição dos testes unitarios e alteração de nomeclaturas

+++ API-Ink-4-You/ink-4-you/src/main/java/br/com/bandtec/ink4yousembanco/controller/EstiloTatuadorController.java
@@ -46,11 +46,12 @@ public class EstiloTatuadorController {
 
     @DeleteMapping(path ={"/{id}"})
     public ResponseEntity <?> deletaEstiloUsuario(@PathVariable Integer id) {
-        return repository.findById(id)
-                .map(record -> {
-                    repository.deleteById(id);
-                    return ResponseEntity.ok().build();
-                }).orElse(ResponseEntity.notFound().build());
+        if (repository.existsById(id)){
+            repository.deleteById(id);
+            return ResponseEntity.status(200).build();
+        }
+
+        return ResponseEntity.status(404).build();
     }
 
 
@@ -104,6 +105,5 @@ public class EstiloTatuadorController {
 
         return ResponseEntity.status(200).body(estiloTatuador);
     }
-
 
 }
